调整地图泡泡距离过长文字的显示

+++ Findlawyer/Findlawyer/Modle/LBSLawfirm.h
@@ -30,7 +30,7 @@
 @property (strong, nonatomic) NSMutableArray  *lawyerist;   // 律师列表
 
 
-@property (assign, nonatomic) NSUInteger distance;                  // 距离
+@property (assign, nonatomic) double distance;                  // 距离
 @property (assign, nonatomic) CLLocationCoordinate2D coordinate;    // 律所坐标
 
 @property (nonatomic, assign) BOOL isShowMapLawfirmAnnotationPaopaoView; // 是否弹出地图上的律所annotation的泡泡视图
